server/requirements.txt
grpcio
grpcio-tools
protobuf
googleapis-common-protos
uvicorn
pydantic
PyJWT
python-dotenv
pydantic-settings
pymongo
asyncio
httpx
requests-cache
retry-requests
openmeteo-requests
pandas
dnspython
pymongo[serv]
pymongo
pymongo[async]
mailtrap